requirements.txt
django
djangorestframework
channels
daphne
djangorestframework-simplejwt
django-cors-headers
mysqlclient
Pillow
whitenoise
